pull samplers into separate file
